pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anomaly-detection"
version = "0.1.0"
description = "Manufacturing Anomaly Detection from IoT Sensor Time Series (NASA CMAPSS)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "uzbekswe"},
]

dependencies = [
    # Data
    "numpy>=1.24,<2.0",
    "pandas>=2.0",
    "scipy>=1.11",

    # ML / Deep Learning
    "torch>=2.1",
    "scikit-learn>=1.3",

    # Experiment Tracking
    "mlflow>=2.9",

    # API Serving
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",

    # Database
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0",

    # Monitoring
    "evidently>=0.4",
    "prometheus-client>=0.19",

    # Dashboard
    "streamlit>=1.29",
    "plotly>=5.18",
    "requests>=2.31",

    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0",

    # HTTP client (for streamlit -> API)
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",

    # Linting & Formatting
    "ruff>=0.1.8",

    # Type Checking
    "mypy>=1.7",

    # Notebooks
    "jupyter>=1.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=src --cov-report=term-missing -v"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
